+++ packages/expo-dev-menu/android/src/debug/java/expo/modules/devmenu/DevMenuFragment.kt
@@ -68,7 +68,8 @@ class DevMenuFragment(
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
 
-    val shouldShowAtLaunch = preferences.showsAtLaunch || !preferences.isOnboardingFinished
+    val shouldShowAtLaunch = !DevMenuSessionOverrides.isAutoLaunchDisabled &&
+      (preferences.showsAtLaunch || !preferences.isOnboardingFinished)
     if (shouldShowAtLaunch) {
       showMenuAtLaunch()
     }